+++ docs/docs/changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.3.0 (in development)
+- Bug fixes for handling of not found places
+
+## v0.2.0 (2025-06-18)
+- Set up GitHub Actions for CI/CD PYPI deployment
 
 ## v0.1.0 (2025-06-18)
 - __Initial release__ of the project to PyPI